+++ wpms-api-server/src/main/java/webpagemanagementsystem/user/controller/UserController.java
@@ -4,7 +4,6 @@ import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -15,20 +14,20 @@ import webpagemanagementsystem.user.exception.DeleteUserException;
 import webpagemanagementsystem.user.exception.DuplicationRegisterException;
 import webpagemanagementsystem.user.exception.NoUseException;
 import webpagemanagementsystem.user.exception.SocialUnauthorizedException;
-import webpagemanagementsystem.user.service.UserService;
+import webpagemanagementsystem.user.service.UserSocialService;
 
 @RestController
 @RequestMapping("/user")
 @RequiredArgsConstructor
 public class UserController {
 
-    private final Map<String, UserService> userServiceMap;
+    private final Map<String, UserSocialService> userSocialServiceMap;
 
 
     @PostMapping("/social/{socialType}")
     public ResponseEntity<String> kakaoLogin(@PathVariable("socialType") String socialType, @RequestBody SocialRequestDto socialRequestDto) {
         try {
-            String accessToken = userServiceMap.get(socialType + "UserService").socialLoginProgress(socialRequestDto.getAccesstoken());
+            String accessToken = userSocialServiceMap.get(socialType + "UserSocialService").socialLoginProgress(socialRequestDto.getAccesstoken());
             return ResponseEntity.ok(accessToken);
         } catch (SocialUnauthorizedException e) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.fillInStackTrace().getMessage());
